refactor(plugin): restructure modules and rename for clarity
- Merge types.rs + call.rs into canister.rs: Candid wire types and
  call wrappers belong together; "call" was too generic a name
- Rename gather.rs -> scan.rs: scanning files is clearer than gathering
- Rename upload.rs -> sync.rs: the module orchestrates a full sync
  (diff, delete, upload), not just uploads
- ContentEncoder -> Encoder (redundant prefix inside content module)
- default_encoders() -> encoders_for() (reads as a query, not a getter)
- AssetDescriptor -> AssetSource, field source -> path
- gather() -> scan(), run() -> sync()
- make_project_asset() -> prepare_asset()
- assemble_operations() -> build_operations()
